refactor: extract useHrefs hook from DownloadAs
